splitting the read and write message queue stage 2
